vs2015: disable warning in generated code, fix another
Macro redefinition is for INT8_MIN, etc. which are generated by flex.
VS doesn't define __STDC_VERSION__ as >= C99 because it's still only
partial support in VS2015.

Fix a float->int conversion narrowing warning.

Reviewed-on: https://chromium-review.googlesource.com/268821
Tested-by: Scott Graham <scottmg@chromium.org>

+++ src/libANGLE/renderer/d3d/d3d9/SwapChain9.cpp
@@ -286,7 +286,7 @@ EGLint SwapChain9::swapRect(EGLint x, EGLint y, EGLint width, EGLint height)
         device->SetStreamSourceFreq(streamIndex, 1);
     }
 
-    D3DVIEWPORT9 viewport = {0, 0, mWidth, mHeight, 0.0f, 1.0f};
+    D3DVIEWPORT9 viewport = {0, 0, static_cast<DWORD>(mWidth), static_cast<DWORD>(mHeight), 0.0f, 1.0f};
     device->SetViewport(&viewport);
 
     float x1 = x - 0.5f;
